+++ internal/pkg/util/shell/escape.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2018, Sylabs, Inc. All rights reserved.
+// Copyright (c) 2018-2021 Sylabs, Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license.  Please
 // consult LICENSE.md file distributed with the sources of this project regarding
 // your rights to use or distribute this software.
@@ -16,7 +16,8 @@ func ArgsQuoted(a []string) (quoted string) {
 	return
 }
 
-// Escape performs escaping of shell quotes, backticks and $ characters
+// Escape performs escaping of shell double quotes, backticks and $ characters.
+// Does not escape single quotes - apply EscapeSingleQuotes separately for this.
 func Escape(s string) string {
 	escaped := strings.Replace(s, `\`, `\\`, -1)
 	escaped = strings.Replace(escaped, `"`, `\"`, -1)
@@ -25,7 +26,12 @@ func Escape(s string) string {
 	return escaped
 }
 
-// EscapeQuotes performs escaping of double quotes only
-func EscapeQuotes(s string) string {
+// EscapeQuotes performs shell escaping of double quotes only
+func EscapeDoubleQuotes(s string) string {
 	return strings.Replace(s, `"`, `\"`, -1)
+}
+
+// EscapeSingleQuotes performs shell escaping of single quotes only
+func EscapeSingleQuotes(s string) string {
+	return strings.Replace(s, `'`, `'"'"'`, -1)
 }